add: replace struct creation with macros
