+++ src/composer/conclusions.rs
@@ -1,13 +1,13 @@
+use crate::ast::program::{DefinitionInserter, Program};
+use crate::ast::agda_expr::AgdaExpr;
 use crate::ast::top_decl::TDeclaration;
 use crate::ast::agda_expr::AgdaExpr::{FunType, QuestionMark};
 use crate::ast::function_type::FunctionType;
-use crate::ast::program::{DefinitionInserter, Program};
 use crate::ast::theorem_decl::Theorem;
-use crate::monty::fresh_variable::to_unicode_subscript;
-use crate::{term, theorem};
-use crate::ast::agda_expr::AgdaExpr;
 use crate::ast::top_decl::TDeclaration::TheoremDecl;
 use crate::ast::agda_expr::AgdaExpr::Term;
+use crate::monty::fresh_variable::to_unicode_subscript;
+use crate::{function_type, term, theorem};
 
 pub fn compose_conclusions(conclusions: Vec<(String, AgdaExpr)>, f: &mut Program) -> () {
 
@@ -15,12 +15,7 @@ pub fn compose_conclusions(conclusions: Vec<(String, AgdaExpr)>, f: &mut Program
     for (conc_name, conc_type) in conclusions {
 
         let iden = format!("thm{}", to_unicode_subscript(assumtion_index));
-
-        let hypothesis = FunType(FunctionType{
-            lhs: term!("KnowledgeBaseᵣ"),
-            rhs: term!(conc_name)
-        });
-
+        let hypothesis = function_type!(*term!("KnowledgeBaseᵣ"), *term!(conc_name));
         let proof = QuestionMark;
 
         let func = theorem!(iden, hypothesis, proof, None);
