+++ src/composer/lambda_to_types.rs
@@ -1,20 +1,14 @@
-use std::collections::HashMap;
-use std::ptr::eq;
-use crate::ccg::rule::CCGRule;
-use crate::lambda::predicate::Predicate;
-use crate::lambda::types::LambdaEntity;
-use crate::lambda::variable::Variable;
-use crate::monty::fresh_variable::to_unicode_subscript;
+use crate::ast::top_decl::TDeclaration;
+use crate::ast::agda_expr::AgdaExpr;
+use crate::ast::program::{DefinitionInserter, Program};
 use crate::ast::application::Application;
 use crate::ast::agda_expr::AgdaExpr::Term;
-use crate::{app, term};
-use crate::ast::agda_expr::AgdaExpr;
 use crate::ast::agda_expr::AgdaExpr::FunType;
 use crate::ast::function_type::FunctionType;
-use crate::ast::program::{DefinitionInserter, Program};
 use crate::ast::record_decl::Record;
 use crate::ast::top_decl::TDeclaration::RecordDecl;
 use crate::ast::var_declaration::VarDecl;
+use crate::{app, function_type, record, term, var_decl};
 use crate::composer::case_converter::*;
 use crate::composer::compose_predicate::compose_predicate;
 use crate::composer::compose_variable::compose_variable;
@@ -24,10 +18,7 @@ pub fn generate_function_header(arity: usize) -> AgdaExpr {
     if arity == 0 {
         *term!("Set")
     } else {
-        FunType(FunctionType{
-            lhs: term!("Entity"),
-            rhs: Box::new(generate_function_header(arity - 1))
-        })
+        function_type!(*term!("Entity"), generate_function_header(arity - 1))
     }
 }
 
@@ -57,27 +48,19 @@ let iden: String = format!("{}×{}", proj1_iden.0, proj2_iden.0)
     .collect();
 
 /* Generate Fields */
-let mut fields: Vec<VarDecl> = vec![
-    VarDecl{
-        iden: "e₁".to_string(),
-        _type: term!(proj1_iden.0)},
-    VarDecl{
-        iden: "e₂".to_string(),
-        _type: term!(proj2_iden.0)},
+let mut fields = vec![
+    var_decl!("e₁", term!(proj1_iden.0)),
+    var_decl!("e₂", term!(proj2_iden.0))
 ];
 
 /* Now, we need to insert the record for it */
 let record_name = format!("{}ᵣ", convert_case(&*iden, CaseStyle::PascalCase));
 let constructor_name = format!("{}꜀", convert_case(&*iden, CaseStyle::PascalCase));
 
-let rec = Record {
-    record_iden: record_name.clone(),
-    constructor_iden: constructor_name,
-    fields: fields,
-    comment: None
-};
 
-f.insert_definition(RecordDecl(rec));
+let rec = record!(record_name, constructor_name, fields, None);
+
+f.insert_definition(rec);
 (record_name, *term!("Temporary"))
 }
