+++ src/composer/synonym_handler.rs
@@ -1,3 +1,5 @@
+use crate::ast::agda_expr::AgdaExpr;
+use crate::ast::program::Program;
 use crate::ast::program::DefinitionInserter;
 use crate::ast::program::PostulateInserter;
 use crate::ast::theorem_decl::Theorem;
@@ -5,16 +7,14 @@ use crate::ast::dependent_function::DependentFunction;
 use crate::ast::application::Application;
 use crate::ast::binary_op::BinOperator;
 use crate::ast::function_type::FunctionType;
-use crate::composer::case_converter::{convert_case, CaseStyle};
-use crate::wordnet::interface::get_meanings;
-use crate::{abstraction, app, bin_op, dependent_function, function_type, term, theorem, var_decl, WORDS_IN_EXISTENCE};
-use crate::ast::agda_expr::AgdaExpr;
 use crate::ast::abstraction::Abstraction;
-use crate::ast::program::Program;
 use crate::ast::var_declaration::VarDecl;
 use crate::ast::agda_expr::AgdaExpr::Term;
 use crate::ast::operator::Operator;
 use crate::ast::top_decl::TDeclaration;
+use crate::composer::case_converter::{convert_case, CaseStyle};
+use crate::wordnet::interface::get_meanings;
+use crate::{abstraction, app, bin_op, dependent_function, function_type, term, theorem, var_decl, WORDS_IN_EXISTENCE};
 /* todo: move these to a higher level when integrating with CLI */
 pub enum SynsetStrategy {
     Ignore, BestMatch, AllMeanings
